requirements.txt
# ── Core RL ───────────────────────────────────────────────────────────────────
gymnasium>=0.29.1
stable-baselines3>=2.3.0
torch>=2.2.0
tensorboard>=2.16.0

# ── Data & ML ─────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=14.0.0
scikit-learn>=1.4.0
scipy>=1.12.0

# ── NLP / Sentiment ───────────────────────────────────────────────────────────
transformers>=4.38.0
newsapi-python>=0.2.7

# ── Visualisation & Dashboard ─────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Utilities ─────────────────────────────────────────────────────────────────
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.7.0
